refactor: probes: move diff probe types to own files
