+++ pkg/ebpf/probes/probes.go
@@ -2,12 +2,8 @@ package probes
 
 import (
 	"fmt"
-	"net"
-	"strings"
-	"syscall"
 
 	bpf "github.com/aquasecurity/libbpfgo"
-	"github.com/aquasecurity/libbpfgo/helpers"
 	"github.com/aquasecurity/tracee/pkg/capabilities"
 	"kernel.org/pub/linux/libs/security/libcap/cap"
 )
@@ -15,28 +11,9 @@ import (
 //
 // Probes:
 //
-
-// when attaching a traceProbe, by handle, to its eBPF program:
-//
-//   Handle == traceProbe (types: rawTracepoint, kprobe, kretprobe)
-//
-//     Attach(EventHandle)
-//     Detach(EventHandle)
-//
-// when attaching a tcProbe, by handle, to its eBPF program:
-//
-//   Handle == tcProbe
-//
-//     Attach(EventHandle, *net.Interface)
-//     Detach(EventHandle, *net.Interface)
-//
-// to detach all probes:
-//
-//     DetachAll()
-//
-// NOTE: keeping both interfaces with variadic args on **purpose** until we
-//       define real use cases, by extending supported "probes" types (trace,
-//       tc, socket, xdp, tunnel, cgroup, ...) **
+// NOTE: keeping both (Probes and Probe) interfaces with variadic args on
+//       **purpose** until we define real use cases, by extending supported
+//       "probes" types (trace, tc, socket, xdp, tunnel, cgroup, ...) **
 //
 
 type Probes interface {
@@ -191,7 +168,7 @@ func (p *probes) Autoload(handle Handle, autoload bool) error {
 }
 
 //
-// probe
+// Probe
 //
 
 type probeType uint8
@@ -210,318 +187,8 @@ type Probe interface {
 }
 
 //
-// traceProbe
+// Event Handles
 //
-
-type traceProbe struct {
-	probeType   probeType
-	eventName   string
-	programName string
-	bpfLink     *bpf.BPFLink
-}
-
-// attach attaches an eBPF program to its probe
-func (p *traceProbe) attach(module *bpf.Module, args ...interface{}) error {
-	var link *bpf.BPFLink
-
-	if p.bpfLink != nil {
-		return nil // already attached, it is ok to call attach again
-	}
-
-	if module == nil {
-		return fmt.Errorf("incorrect arguments for event: %s", p.eventName)
-	}
-
-	prog, err := module.GetProgram(p.programName)
-	if err != nil {
-		return err
-	}
-
-	switch p.probeType {
-	case kprobe:
-		link, err = prog.AttachKprobe(p.eventName)
-	case kretprobe:
-		link, err = prog.AttachKretprobe(p.eventName)
-	case tracepoint:
-		tp := strings.Split(p.eventName, ":")
-		tpClass := tp[0]
-		tpEvent := tp[1]
-		link, err = prog.AttachTracepoint(tpClass, tpEvent)
-	case rawTracepoint:
-		tpEvent := strings.Split(p.eventName, ":")[1]
-		link, err = prog.AttachRawTracepoint(tpEvent)
-	}
-
-	if err != nil {
-		return fmt.Errorf("failed to attach event: %s (%v)", p.eventName, err)
-	}
-
-	p.bpfLink = link
-
-	return nil
-}
-
-// detach detaches an eBPF program from its probe
-func (p *traceProbe) detach(args ...interface{}) error {
-	var err error
-
-	if p.bpfLink == nil {
-		return nil // already detached, it is ok to call detach again
-	}
-
-	err = p.bpfLink.Destroy()
-	if err != nil {
-		return fmt.Errorf("failed to detach event: %s (%v)", p.eventName, err)
-	}
-
-	p.bpfLink = nil // NOTE: needed so a new call to bpf_link__destroy() works
-
-	return nil
-}
-
-// autoload sets an eBPF program to autoload (true|false)
-func (p *traceProbe) autoload(module *bpf.Module, autoload bool) error {
-	return enableDisableAutoload(module, p.programName, autoload)
-}
-
-//
-// uProbe
-//
-
-type uProbe struct {
-	eventName   string
-	programName string // eBPF program to execute when uprobe triggered
-	binaryPath  string // ELF file path to attach uprobe to
-	symbolName  string // ELF binary symbol to attach uprobe to
-	bpfLink     *bpf.BPFLink
-}
-
-// attach attaches an eBPF program to its probe
-func (p *uProbe) attach(module *bpf.Module, args ...interface{}) error {
-	var link *bpf.BPFLink
-
-	if p.bpfLink != nil {
-		return nil // already attached, it is ok to call attach again
-	}
-
-	if module == nil {
-		return fmt.Errorf("incorrect arguments for event: %s", p.eventName)
-	}
-
-	prog, err := module.GetProgram(p.programName)
-	if err != nil {
-		return err
-	}
-
-	offset, err := helpers.SymbolToOffset(p.binaryPath, p.symbolName)
-	if err != nil {
-		return fmt.Errorf("error finding %s function offset: %v", p.symbolName, err)
-	}
-
-	link, err = prog.AttachUprobe(-1, p.binaryPath, offset)
-	if err != nil {
-		return fmt.Errorf("error attaching uprobe on %s: %v", p.symbolName, err)
-	}
-
-	p.bpfLink = link
-
-	return nil
-}
-
-// detach detaches an eBPF program from its probe
-func (p *uProbe) detach(args ...interface{}) error {
-	var err error
-
-	if p.bpfLink == nil {
-		return nil // already detached, it is ok to call detach again
-	}
-
-	err = p.bpfLink.Destroy()
-	if err != nil {
-		return fmt.Errorf("failed to detach event: %s (%v)", p.eventName, err)
-	}
-
-	p.bpfLink = nil // NOTE: needed so a new call to bpf_link__destroy() works
-
-	return nil
-}
-
-// autoload sets an eBPF program to autoload (true|false)
-func (p *uProbe) autoload(module *bpf.Module, autoload bool) error {
-	return enableDisableAutoload(module, p.programName, autoload)
-}
-
-//
-// tcProbe
-//
-
-type tcProbe struct {
-	programName   string
-	tcHooks       map[*net.Interface]*bpf.TcHook
-	tcAttachPoint bpf.TcAttachPoint
-	skipLoopback  bool
-}
-
-// attach attaches an eBPF program to its probe
-func (p *tcProbe) attach(module *bpf.Module, args ...interface{}) error {
-	return p.attachOrDetach(module, args...)
-}
-
-// detach detaches an eBPF program from its probe
-func (p *tcProbe) detach(args ...interface{}) error {
-	return p.attachOrDetach(nil, args...)
-}
-
-// attachOrDetach resolves variadic argument and calls appropriate function
-func (p *tcProbe) attachOrDetach(module *bpf.Module, args ...interface{}) error {
-	var netIface *net.Interface
-
-	for _, arg := range args {
-		switch a := arg.(type) {
-		case *net.Interface:
-			netIface = a
-		}
-	}
-
-	if module != nil {
-		return p.attachTc(module, netIface)
-	}
-
-	if netIface == nil { // detach from all if no interface given
-		return p.detachTcAll(module)
-	}
-
-	return p.detachTc(module, netIface)
-}
-
-// attachTc attaches a tc program to a given interface
-func (p *tcProbe) attachTc(module *bpf.Module, netIface *net.Interface) error {
-	var err error
-
-	if netIface == nil {
-		return fmt.Errorf("missing interface to attach to: %s", p.programName)
-	}
-
-	// loopback devices are special, some tc probes should be skipped
-	isNetIfaceLo := netIface.Flags&net.FlagLoopback == net.FlagLoopback
-	if isNetIfaceLo && p.skipLoopback {
-		return nil
-	}
-
-	if _, ok := p.tcHooks[netIface]; ok {
-		return fmt.Errorf("%s already attached to %s", p.programName, netIface.Name)
-	}
-
-	hook := module.TcHookInit()
-	if hook == nil {
-		return fmt.Errorf("could not initialize TcHook %s", p.programName)
-	}
-
-	err = hook.SetInterfaceByName(netIface.Name)
-	if err != nil {
-		return fmt.Errorf("failed to set tc hook interface: %v", err)
-	}
-
-	hook.SetAttachPoint(p.tcAttachPoint)
-
-	err = hook.Create()
-	if err != nil {
-		if errno, ok := err.(syscall.Errno); ok && errno != syscall.EEXIST {
-			return fmt.Errorf("tc hook create: %v", err)
-		}
-	}
-
-	prog, _ := module.GetProgram(p.programName)
-	if prog == nil {
-		return fmt.Errorf("could not find tc program %s", p.programName)
-	}
-
-	tcOpts := bpf.TcOpts{ProgFd: int(prog.FileDescriptor())}
-
-	err = hook.Attach(&tcOpts)
-	if err != nil {
-		return fmt.Errorf("tc attach: %v", err)
-	}
-
-	if p.tcHooks == nil {
-		p.tcHooks = make(map[*net.Interface]*bpf.TcHook)
-	}
-
-	p.tcHooks[netIface] = hook
-
-	return err
-}
-
-// detachTc detaches a tc program from a given interface it is attached to
-func (p *tcProbe) detachTc(module *bpf.Module, netIface *net.Interface) error {
-	var err error
-
-	if _, ok := p.tcHooks[netIface]; !ok {
-		return fmt.Errorf("%s not attached to %s", p.programName, netIface.Name)
-	}
-
-	hook := p.tcHooks[netIface]
-
-	// TODO: https://github.com/aquasecurity/tracee/issues/1828
-	//
-	// 1. Entire clsact qdisc is purged when tc hook is destroyed:
-	//    it might destroy existing filters (other eBPF programs)
-	//    unrelated to tracee.
-	//
-	// 2. eBPF TC detach does not work (destroy needs to be used)
-
-	if hook != nil {
-		err = hook.Destroy()
-		if err != nil {
-			return err
-		}
-	}
-
-	delete(p.tcHooks, netIface)
-
-	return err
-}
-
-// detachTcAll detaches a tc program from all interfaces it is attached to
-func (p *tcProbe) detachTcAll(module *bpf.Module) error {
-	var err error
-
-	for netIface := range p.tcHooks {
-		err = p.detachTc(module, netIface)
-		if err != nil {
-			return err
-		}
-	}
-
-	return err
-}
-
-// autoload sets an eBPF program to autoload (true|false)
-func (p *tcProbe) autoload(module *bpf.Module, autoload bool) error {
-	return enableDisableAutoload(module, p.programName, autoload)
-}
-
-//
-// common function(s) to Probe implementations
-//
-
-// enableDisableAutoload enables or disables an eBPF program autoload setting
-func enableDisableAutoload(module *bpf.Module, programName string, autoload bool) error {
-	var err error
-
-	if module == nil || programName == "" {
-		return fmt.Errorf("incorrect arguments (program: %s)", programName)
-	}
-
-	prog, err := module.GetProgram(programName)
-	if err != nil {
-		return err
-	}
-
-	return prog.SetAutoload(autoload)
-}
-
-// event handles
 
 type Handle int32
